iOS 3 Compatibility Attempt

+++ platforms/ios/EAGLView.h
@@ -5,8 +5,6 @@
 
 #import <UIKit/UIKit.h>
 
-#import <OpenGLES/ES1/gl.h>
-#import <OpenGLES/ES1/glext.h>
 #import <OpenGLES/ES2/gl.h>
 #import <OpenGLES/ES2/glext.h>
 
@@ -26,7 +24,7 @@
     GLuint defaultFramebuffer, colorRenderbuffer;
 	GLuint depthRenderbuffer;
     
-    float viewScale;
+    CGFloat viewScale;
 }
 
 @property (nonatomic, retain) EAGLContext *context;
